Revert "Add build flag for retrier implements"
This reverts commit 518a7a785cd818ab525b642fc72f5aa3a3fa6b36.
